// src/app/model/user.ts
export class User {
    Username: string;
    Password: string;
    Profiles: Array<string> = [];
    Systems: Array<string> = [];
    StandardCost: number;
    StandardSchedule: string;
    DateOfBirth: string;
    Description: string;
    OvertimeCost: number;
    OvertimeSchedule: string;
    Skill: string;
    FirstName: string;
    LastName: string;
    Email: string;
    Telephone: string;
    Organization: string;
    Group: string;
    Note: string;
    Tag: string;
    Timestamp: string;
    Street: string;
    Location: string;
    City: string;
    State: string;
    Cap: string;
    Language: string;
}